+++ src/spoke/instances/SpokeInstance.sol
@@ -18,8 +18,8 @@ contract SpokeInstance is Spoke {
   /// @param maxAllowedBorrowedReserves_ The maximum allowed number of borrowed reserves per user.
   constructor(
     address oracle_,
-    uint8 maxAllowedCollateralReserves_,
-    uint8 maxAllowedBorrowedReserves_
+    uint64 maxAllowedCollateralReserves_,
+    uint64 maxAllowedBorrowedReserves_
   ) Spoke(oracle_, maxAllowedCollateralReserves_, maxAllowedBorrowedReserves_) {
     _disableInitializers();
   }
