+++ tests/mocks/MockSpoke.sol
@@ -26,8 +26,8 @@ contract MockSpoke is Spoke, Test {
 
   constructor(
     address oracle_,
-    uint8 maxAllowedCollateralReserves_,
-    uint8 maxAllowedBorrowedReserves_
+    uint64 maxAllowedCollateralReserves_,
+    uint64 maxAllowedBorrowedReserves_
   ) Spoke(oracle_, maxAllowedCollateralReserves_, maxAllowedBorrowedReserves_) {}
 
   function initialize(address) external override {}
